requirements.txt
appdirs==1.4.4
black==21.7b0
click==8.0.1
joblib==1.0.1
mypy-extensions==0.4.3
nltk==3.6.2
pathspec==0.9.0
PyStemmer==2.0.1
regex==2021.8.28
tomli==1.2.1
tqdm==4.62.2
